backend/ai/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
torch==2.1.0
transformers==4.30.0
huggingface-hub==0.19.4
